simple_chatbot/pyproject.toml
[project]
name = "simple-chatbot"
version = "0.1.0"
description = "A simple project to demonstrate how to use LLMs as chatbots"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.76.0",
    "python-dotenv>=1.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
